DateSort version
The DateSort version shows how we would sort items in a 2D Array by date. There is some complexity as we have to write the date as a string, yet when we sort, we need to sort a datetime object. This requires a conversion using the strptime method. Once sorted, we can convert these dates back to strings using the strftime method and output our sorted 2D array.